# cmd/contentmanagement/docs/swagger.yaml
basePath: /api/v1
definitions:
  content.Content:
    properties:
      contentDefinitionID:
        type: string
      created:
        type: string
      data:
        $ref: '#/definitions/content.ContentData'
      id:
        type: string
      updated:
        type: string
    type: object
  content.ContentData:
    properties:
      contentID:
        type: string
      created:
        description: 'TODO: does ContentData need a Created Field?'
        type: string
      properties:
        $ref: '#/definitions/content.ContentLanguage'
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      version:
        type: integer
    type: object
  content.ContentField:
    properties:
      id:
        type: string
      localized:
        type: boolean
      type:
        type: string
      value: {}
    type: object
  content.ContentFields:
    additionalProperties:
      $ref: '#/definitions/content.ContentField'
    type: object
  content.ContentLanguage:
    additionalProperties:
      $ref: '#/definitions/content.ContentFields'
    type: object
  content.CreateContentRequest:
    properties:
      contentDefinitionId:
        type: string
    type: object
  content.OKResult:
    type: object
  content.UpdateContentRequestBody:
    properties:
      fields:
        additionalProperties: true
        description: Properties
        type: object
      language:
        description: Language
        type: string
      version:
        description: Version
        type: integer
    type: object
...
  models.ErrorBody:
    properties:
      fieldName:
        type: string
      message:
        description: 'required: true'
        type: string
    type: object
  models.GenericError:
    properties:
      body:
        $ref: '#/definitions/models.ErrorBody'
        description: 'in: body'
      statusCode:
        description: swagger:ignore
        type: integer
    type: object
  models.OKResult:
    type: object
  query.ContentListReadModel:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /contentmanagement/content:
    get:
      consumes:
      - application/json
      description: list all content
      parameters:
      - description: uuid formatted ID.
        format: uuid
        in: query
        items:
          type: string
        name: cid
        required: true
        type: array
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/query.ContentListReadModel'
            type: array
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.GenericError'
      summary: List all content
      tags:
      - content
    post:
      consumes:
      - application/json
      description: Creates new content basen on a contentdefinition
      parameters:
      - description: contentdefinitionid
        in: body
        name: contentdefinitionid
        required: true
        schema:
          $ref: '#/definitions/content.CreateContentRequest'
      responses:
        "201":
          description: Created
          headers:
            Location:
              type: string
          schema:
            $ref: '#/definitions/content.Content'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.GenericError'
      summary: Create new content
      tags:
      - content
  /contentmanagement/content/{id}:
    delete:
      consumes:
      - application/json
      description: Archives content with ID
      parameters:
      - description: uuid formatted ID.
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/content.OKResult'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.GenericError'
      summary: Archives content
      tags:
      - content
    get:
      consumes:
      - application/json
      description: Get content by id and optionally version
      parameters:
      - description: uuid formatted ID.
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: content version
        in: query
        name: version
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/content.Content'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.GenericError'
      summary: Get content by id
      tags:
      - content
    put:
      consumes:
      - application/json
      description: Update content
      parameters:
      - description: uuid formatted ID.
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: body
        in: body
        name: requestbody
        required: true
        schema:
          $ref: '#/definitions/content.UpdateContentRequestBody'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OKResult'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.GenericError'
      summary: Update content
      tags:
      - content
  /contentmanagement/content/{id}/publish:
    post:
      consumes:
      - application/json
      description: Publishes content with ID
      parameters:
      - description: uuid formatted ID.
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: content version
        in: query
        name: version
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/content.OKResult'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.GenericError'
      summary: Publishes content
      tags:
      - content
...
swagger: "2.0"